requirements.txt
numpy
jupyter
matplotlib
scikit-image
torchvision
torchaudio
